Add files via upload
This program uses multiple functions and helper functions to calculate
and return different calculations within an options menu. It allows the user
to generate a secure password, calculate percentages from a fraction,
calculate the days between the current date and july, 4th, 2025, calculate the law
of cosine of a triangle,and calculate the volume of a right cylinder.